+++ test/TexasHoldemRoom3pRealKeysShuffledOnce256bit.t.sol
@@ -7,7 +7,7 @@ import "../src/PokerHandEvaluatorv2.sol";
 import "../src/BigNumbers/BigNumbers.sol";
 import "../src/CryptoUtils.sol";
 
-// REQUIRED: in TexasHoldemRoom.sol, change uint8 public constant MIN_PLAYERS = 3;
+// REQUIRED: in TexasHoldemRoom.sol, change uint256 public constant MIN_PLAYERS = 3;
 
 contract TexasHoldemRoom3pRealKeysShuffledOnceTest is Test {
     using BigNumbers for BigNumber;
@@ -484,7 +484,7 @@ contract TexasHoldemRoom3pRealKeysShuffledOnceTest is Test {
     //         uint256 totalRoundBet,
     //         bool hasFolded,
     //         bool isAllIn,
-    //         uint8 seatPosition
+    //         uint256 seatPosition
     //     ) = room.players(0);
     //     assertEq(addr, player1);
     //     assertEq(chips, INITIAL_BALANCE);
@@ -611,7 +611,7 @@ contract TexasHoldemRoom3pRealKeysShuffledOnceTest is Test {
         // the encrypted values submitted are not correct at the moment
         // submit card index and decrypted value for each card
         // card 0 to p1, card 1 to p2, card 2 to p1, card 3 to p2!
-        uint8[] memory cardIndexes = new uint8[](4);
+        uint256[] memory cardIndexes = new uint256[](4);
         BigNumber[] memory decryptionValues = new BigNumber[](4);
         cardIndexes[0] = 1;
         cardIndexes[1] = 2;
@@ -751,7 +751,7 @@ contract TexasHoldemRoom3pRealKeysShuffledOnceTest is Test {
         console.log("Reveal flop stage reached");
 
         // All players should submit their decryption values for the flop cards
-        uint8[] memory cardIndexesFlop = new uint8[](3);
+        uint256[] memory cardIndexesFlop = new uint256[](3);
         BigNumber[] memory decryptionValuesFlop = new BigNumber[](3);
         // player 1 should submit their decryption values for flop cards
         vm.startPrank(player1);
@@ -880,7 +880,7 @@ contract TexasHoldemRoom3pRealKeysShuffledOnceTest is Test {
         assertEq(uint256(room.stage()), uint256(TexasHoldemRoom.GameStage.RevealTurn));
 
         // ======= TURN CARD REVEAL =======
-        uint8[] memory cardIndexesTurn = new uint8[](1);
+        uint256[] memory cardIndexesTurn = new uint256[](1);
         BigNumber[] memory decryptionValuesTurn = new BigNumber[](1);
         // 3 players, 6th card is burned, 789 are the flop cards,
         // 10th card is burned, 11th card is the turn card
@@ -947,7 +947,7 @@ contract TexasHoldemRoom3pRealKeysShuffledOnceTest is Test {
         // ======= START OF RIVER CARD REVEAL =======
         assertEq(uint256(room.stage()), uint256(TexasHoldemRoom.GameStage.RevealRiver));
         // All players should submit their decryption values for the river card
-        uint8[] memory cardIndexesRiver = new uint8[](1);
+        uint256[] memory cardIndexesRiver = new uint256[](1);
         BigNumber[] memory decryptionValuesRiver = new BigNumber[](1);
         // 3 players, 10th card is burned, 11th card is the turn card
         // 12th card is burned, 13th card is the river card
@@ -1081,7 +1081,7 @@ contract TexasHoldemRoom3pRealKeysShuffledOnceTest is Test {
         );
         // Player 3 won!
         vm.expectEmit(address(room));
-        uint8[] memory expectedWinnerPlayerIndicies = new uint8[](1);
+        uint256[] memory expectedWinnerPlayerIndicies = new uint256[](1);
         expectedWinnerPlayerIndicies[0] = 2;
 
         address[] memory winnerAddresses = new address[](1);
